Avoid checking more recent version constants in systest
This fixes CI and avoids needing a bleeding edge jni.h

This adds jni19,jni20 and jni21 feature flags for systest so we can skip
checking for the corresponding JNI_VERSION constants unless we know
we have a suitably recent version of Java installed.

CI will now install temurin v20 so it can enable the jni19 and jni20
features.